move renderSharedTemplate helper to respoke-style for more sharing
Originally the helper used the dist flag from the gulp file which
required wrapping in a function to pass it when requiring respoke-style.
This turned out to not be needed as the top level pretty setting for
jade was respected by the renderer helper. I left the more idiomatic
function wrapper for future complexity or a reintroduction of the need
to pass values when requiring respoke-style and it's helpers.